Erros de Rotas corrigidos
