+++ front/assets/adaptive-icon.png
