+++ front/package.json
@@ -1,45 +1,28 @@
 {
-  "name": "front",
-  "main": "expo-router/entry",
+  "name": "vendaatacado",
   "version": "1.0.0",
+  "main": "index.ts",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
     "ios": "expo start --ios",
-    "web": "expo start --web",
-    "test": "jest --watchAll"
-  },
-  "jest": {
-    "preset": "jest-expo"
+    "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/vector-icons": "^14.0.2",
     "@react-native-async-storage/async-storage": "^2.1.2",
-    "@react-navigation/native": "^7.0.14",
-    "expo": "~52.0.39",
-    "expo-font": "~13.0.4",
-    "expo-linking": "~7.0.5",
-    "expo-router": "~4.0.19",
-    "expo-splash-screen": "~0.29.22",
+    "@react-navigation/bottom-tabs": "^7.3.10",
+    "@react-navigation/native": "^7.1.6",
+    "expo": "~52.0.43",
     "expo-status-bar": "~2.0.1",
-    "expo-system-ui": "~4.0.8",
-    "expo-web-browser": "~14.0.2",
     "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "react-native": "0.76.7",
-    "react-native-reanimated": "~3.16.1",
+    "react-native": "0.76.9",
     "react-native-safe-area-context": "4.12.0",
-    "react-native-screens": "~4.4.0",
-    "react-native-web": "~0.19.13"
+    "react-native-screens": "~4.4.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
     "@types/react": "~18.3.12",
-    "@types/react-native": "^0.72.8",
-    "jest": "^29.2.1",
-    "jest-expo": "~52.0.6",
-    "react-test-renderer": "18.3.1",
-    "typescript": "~5.3.3"
+    "typescript": "^5.3.3"
   },
   "private": true
 }
