+++ front/routes/lojaNavigator.tsx
@@ -2,15 +2,15 @@ import React from 'react'
 import { createBottomTabNavigator } from  '@react-navigation/bottom-tabs'
 import { Ionicons } from '@expo/vector-icons'
 
-import Perfil from '@/app/screens/loja/perfil'
-import Pedidos from '@/app/screens/loja/pedidos'
-import Estoque from '@/app/screens/loja/estoque'
+import Perfil from '../screens/loja/perfil';
+import Pedidos from '../screens/loja/pedidos';
+import Estoque from '../screens/loja/estoque';
 
 const Tab = createBottomTabNavigator();
 
 export default function LojaNavigator(){
     return(
-        <Tab.Navigator screenOptions={({ route }) => ({
+        <Tab.Navigator screenOptions={({ route }) => ({ headerShown: false,
             tabBarIcon: ({color,size}) => {
                 let iconName: keyof typeof Ionicons.glyphMap;
 
@@ -30,9 +30,9 @@ export default function LojaNavigator(){
                 return <Ionicons name={iconName} size={size} color={color} />
             },
         })}>
-            <Tab.Screen name='Perfil' component={Perfil} />
             <Tab.Screen name='Pedidos' component={Pedidos} />
             <Tab.Screen name='Estoque' component={Estoque} />
+            <Tab.Screen name='Perfil' component={Perfil} />
         </Tab.Navigator>
     )
 }
