+++ front/screens/loja/estoque.tsx
